moved BinnedDataFitter to Fitters
