+++ mermithid/processors/Fitters/__init__.py
@@ -1,0 +1,7 @@
+'''
+'''
+
+from __future__ import absolute_import
+
+from .BinnedDataFitter import BinnedDataFitter
+
