+++ mermithid/processors/__init__.py
@@ -7,3 +7,4 @@ from . import IO
 from . import TritiumSpectrum
 from . import plots
 from . import misc
+from . import Fitters
