+++ mermithid/processors/misc/__init__.py
@@ -5,4 +5,3 @@ from __future__ import absolute_import
 
 from .FrequencyEnergyConversionProcessor import FrequencyEnergyConversionProcessor
 from .FrequencyShifter import FrequencyShifter
-from .BinnedDataFitter import BinnedDataFitter
